+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleased]
+
+## [v2.1.0]- 2020-12-24
+### Fixed
+
+- FIX([62](https://github.com/meateam/user-service/pull/62)): fix GetApproverInfo with rabaz integration.
+
+### Added
+
+- FEAT([75](https://github.com/meateam/user-service/pull/62)): new RPC method CanApproveToUser
+
 
 ## [v2.0.0] - 2020-10-28
 
@@ -13,9 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- FIX([53](https://github.com/meateam/user-service/pull/53)): fix use undefind lastName and convert it to empty string. 
+- FIX([53](https://github.com/meateam/user-service/pull/53)): fix use undefind lastName and convert it to empty string.
 
-## [Unreleased]
 
 [unreleased]: https://github.com/meateam/user-service/compare/master...develop
 [v2.0.0]: https://github.com/meateam/user-service/compare/v1.3...v2.0.0
+[v2.1.0]: https://github.com/meateam/user-service/compare/v2.0.0...v2.1.0